requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
httpx>=0.27.0,<1.0.0
jinja2>=3.1.4,<4.0.0
python-pptx>=1.0.2,<2.0.0
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
zstandard>=0.23.0,<1.0.0
